backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.2.0
scikit-learn==1.4.1
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9